# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tıraş - Barber Appointment SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
  /account:
    {
      "get":
        {
          "operationId": "getAccount",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties":
                                {
                                  "name": { "type": "string" },
                                  "email": { "type": "string" },
                                  "phone":
                                    { "type": "string", "nullable": true },
                                  "avatar":
                                    { "type": "string", "nullable": true },
                                  "gallery":
                                    {
                                      "type": "array",
                                      "items": { "type": "string" },
                                    },
                                  "emailVerifiedAt":
                                    { "type": "string", "nullable": true },
                                  "phoneVerifiedAt":
                                    { "type": "string", "nullable": true },
                                },
                              "required":
                                [
                                  "name",
                                  "email",
                                  "phone",
                                  "avatar",
                                  "gallery",
                                  "emailVerifiedAt",
                                  "phoneVerifiedAt",
                                ],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
      "patch":
        {
          "operationId": "updateAccount",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "name": { "type": "string" },
                              "email": { "type": "string" },
                              "phone": { "type": "string" },
                              "currentPassword": { "type": "string" },
                            },
                          "required":
                            ["name", "email", "phone", "currentPassword"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "ok": { "type": "boolean" } },
                              "required": ["ok"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
      "delete":
        {
          "operationId": "deleteAccount",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "currentPassword": { "type": "string" },
                              "confirmation": { "type": "string" },
                            },
                          "required": ["currentPassword", "confirmation"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "ok": { "type": "boolean" } },
                              "required": ["ok"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /account/media:
    {
      "put":
        {
          "operationId": "saveAccountMedia",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "avatar": { "type": "string", "nullable": true },
                              "gallery":
                                {
                                  "type": "array",
                                  "items": { "type": "string" },
                                },
                            },
                          "required": ["avatar", "gallery"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "ok": { "type": "boolean" } },
                              "required": ["ok"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /barbers/{id}/media:
    {
      "get":
        {
          "operationId": "getBarberMedia",
          "parameters":
            [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "integer" },
              },
            ],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties":
                                {
                                  "avatar":
                                    { "type": "string", "nullable": true },
                                  "gallery":
                                    {
                                      "type": "array",
                                      "items": { "type": "string" },
                                    },
                                },
                              "required": ["avatar", "gallery"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /account/password:
    {
      "post":
        {
          "operationId": "changeAccountPassword",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "currentPassword": { "type": "string" },
                              "newPassword": { "type": "string" },
                            },
                          "required": ["currentPassword", "newPassword"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "ok": { "type": "boolean" } },
                              "required": ["ok"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /account/logout-all:
    {
      "post":
        {
          "operationId": "logoutAllSessions",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "ok": { "type": "boolean" } },
                              "required": ["ok"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /account/sessions:
    {
      "get":
        {
          "operationId": "listAccountSessions",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "array",
                              "items":
                                {
                                  "type": "object",
                                  "properties":
                                    {
                                      "id": { "type": "string" },
                                      "device": { "type": "string" },
                                      "lastSeenAt": { "type": "string" },
                                      "current": { "type": "boolean" },
                                    },
                                  "required":
                                    ["id", "device", "lastSeenAt", "current"],
                                },
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /account/sessions/{id}:
    {
      "delete":
        {
          "operationId": "revokeAccountSession",
          "security": [{ "bearerAuth": [] }],
          "parameters":
            [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "string" },
              },
            ],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "ok": { "type": "boolean" } },
                              "required": ["ok"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /account/verification:
    {
      "post":
        {
          "operationId": "requestContactVerification",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "channel":
                                {
                                  "type": "string",
                                  "enum": ["email", "phone"],
                                },
                            },
                          "required": ["channel"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "id": { "type": "integer" } },
                              "required": ["id"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
  /account/verification/confirm:
    {
      "post":
        {
          "operationId": "confirmContactVerification",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "id": { "type": "integer" },
                              "code": { "type": "string" },
                            },
                          "required": ["id", "code"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "ok": { "type": "boolean" } },
                              "required": ["ok"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
              "503": { "description": "Delivery unavailable" },
            },
        },
    }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...